+++ android/app/src/main/java/poke/rogue/helper/presentation/biome/BiomeActivity.kt
@@ -12,6 +12,7 @@ import poke.rogue.helper.databinding.ActivityBiomeBinding
 import poke.rogue.helper.presentation.base.error.ErrorHandleActivity
 import poke.rogue.helper.presentation.base.error.ErrorHandleViewModel
 import poke.rogue.helper.presentation.biome.detail.BiomeDetailActivity
+import poke.rogue.helper.presentation.biome.guide.BiomeGuideActivity
 import poke.rogue.helper.presentation.biome.model.toUi
 import poke.rogue.helper.presentation.util.context.startActivity
 import poke.rogue.helper.presentation.util.logClickEvent
